refactor: move Forge surface to Game Director seam (#140)
