+++ application/src/GameDirector.Mcp/Forge/ForgeAuthenticationMiddleware.cs
