+++ application/src/GameDirector.Mcp/Forge/ForgeMapFetchUrlValidator.cs
