+++ application/src/GameDirector.Mcp/Forge/ForgeSessionOptions.cs
