+++ application/src/GameDirector.Mcp/Forge/ForgeTenantMiddleware.cs
