+++ application/src/GameDirector.Mcp/Forge/ForgeTenantNormalizer.cs
