+++ application/src/GameDirector.Mcp/Forge/IForgeStateService.cs
