+++ application/src/GameDirector.Mcp/Forge/InMemoryForgeStateService.cs
