+++ application/src/GameDirector.Mcp/Forge/ModelAugmentedMaterialIntelligenceService.cs
