+++ application/src/GameDirector.Mcp/Forge/ModelAugmentedVectorMapIntelligenceService.cs
