+++ application/src/GameDirector.Mcp/GameDirector.Mcp.csproj
@@ -6,8 +6,13 @@
     <RootNamespace>GameDirector.Mcp</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <InternalsVisibleTo Include="Nexo.Tests.GameDirector" />
+  </ItemGroup>
+  <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="LiteDB" />
     <ProjectReference Include="..\GameDirector.Domain\GameDirector.Domain.csproj" />
+    <ProjectReference Include="..\..\..\src\Nexo.Abstractions\Nexo.Abstractions.csproj" />
     <ProjectReference Include="..\..\..\src\Nexo.Client\Nexo.Client.csproj" />
     <ProjectReference Include="..\..\..\src\Nexo.Core.Application\Nexo.Core.Application.csproj" />
     <ProjectReference Include="..\..\..\src\Nexo.Core.Domain\Nexo.Core.Domain.csproj" />
