+++ application/src/Nexo.API/Nexo.API.csproj
@@ -25,7 +25,6 @@
   <ItemGroup>
     <PackageReference Include="MediatR" />
     <PackageReference Include="Swashbuckle.AspNetCore" />
-    <PackageReference Include="LiteDB" />
     <ProjectReference Include="../../../src/Nexo.BackgroundAgents.HostRunners/Nexo.BackgroundAgents.HostRunners.csproj" />
     <ProjectReference Include="../../../src/Nexo.Brick.Contracts/Nexo.Brick.Contracts.csproj" />
     <ProjectReference Include="../../../src/Nexo.Contracts/Nexo.Contracts.csproj" />
@@ -33,7 +32,6 @@
     <ProjectReference Include="../../../src/Nexo.Ingress.DynamoDb/Nexo.Ingress.DynamoDb.csproj" />
     <ProjectReference Include="../../../src/Nexo.Ingress.AwsSns/Nexo.Ingress.AwsSns.csproj" />
     <ProjectReference Include="../../../src/Nexo.Infrastructure/Nexo.Infrastructure.csproj" />
-    <ProjectReference Include="..\Nexo.GameDomain\Nexo.GameDomain.csproj" />
   </ItemGroup>
 
 </Project>
