+++ application/src/Nexo.Tests.CLI/Nexo.Tests.CLI.csproj
@@ -25,7 +25,6 @@
     <ProjectReference Include="../../../src/Nexo.Core.Application/Nexo.Core.Application.csproj" />
     <ProjectReference Include="..\Nexo.CLI\Nexo.CLI.csproj" />
     <ProjectReference Include="../../../src/Nexo.Infrastructure/Nexo.Infrastructure.csproj" />
-    <ProjectReference Include="..\Nexo.GameDomain\Nexo.GameDomain.csproj" />
   </ItemGroup>
 
 </Project>
