+++ application/src/Nexo.Tests.GameDomain/AssetDescriptorTests.cs
@@ -76,6 +76,29 @@ public class AssetDescriptorTests
     }
 
     [Fact]
+    public void MaterialDescriptor_JsonRoundTrip_PreservesRenderMode()
+    {
+        var original = new MaterialDescriptor
+        {
+            Id = "mat-1",
+            Name = "Test Mat",
+            ShaderName = "Standard",
+            Color = "#FF0000",
+            Metallic = 0.5,
+            Smoothness = 0.7,
+            RenderMode = "Transparent"
+        };
+
+        var opts = new JsonSerializerOptions(JsonSerializerDefaults.Web) { PropertyNameCaseInsensitive = true };
+        var json = JsonSerializer.Serialize(original, opts);
+        var restored = JsonSerializer.Deserialize<MaterialDescriptor>(json, opts);
+
+        restored.Should().NotBeNull();
+        restored!.Id.Should().Be("mat-1");
+        restored.RenderMode.Should().Be("Transparent");
+    }
+
+    [Fact]
     public void PrefabDescriptor_DefaultValues()
     {
         var prefab = new PrefabDescriptor();
@@ -190,6 +213,42 @@ public class AssetDescriptorTests
         audio.BypassReverb.Should().BeFalse();
         audio.Variations.Should().BeEmpty();
         audio.Tags.Should().BeEmpty();
+    }
+
+    [Fact]
+    public void AudioDescriptor_JsonRoundTrip_PreservesSpatialAndVariationMetadata()
+    {
+        var original = new AudioDescriptor
+        {
+            Id = "explosion",
+            Name = "Explosion SFX",
+            Category = "sfx",
+            SubCategory = "explosion",
+            Volume = 0.8,
+            Pitch = 0.9,
+            PitchVariance = 0.1,
+            SpatialBlend = 1.0,
+            Loop = false,
+            MixerGroup = "SFX",
+            Variations =
+            [
+                new AudioVariation { VariantId = "v1", PitchOffset = 0.05, ClipSuffix = "_01" }
+            ],
+            Tags = ["combat", "explosion"]
+        };
+
+        var opts = new JsonSerializerOptions(JsonSerializerDefaults.Web) { PropertyNameCaseInsensitive = true };
+        var json = JsonSerializer.Serialize(original, opts);
+        var restored = JsonSerializer.Deserialize<AudioDescriptor>(json, opts);
+
+        restored.Should().NotBeNull();
+        restored!.Id.Should().Be("explosion");
+        restored.SubCategory.Should().Be("explosion");
+        restored.SpatialBlend.Should().Be(1.0);
+        restored.MixerGroup.Should().Be("SFX");
+        restored.Variations.Should().HaveCount(1);
+        restored.Variations[0].VariantId.Should().Be("v1");
+        restored.Tags.Should().HaveCount(2);
     }
 
     [Fact]
